Major particle filter improvements
I know have a skeleton of how the code for the particle filter should work. There are some math issues to be worked out and it needs to be improved but this  follows the workflow needed for a successfull resampling particle filter. Most changes were altering the initial forecast function to predict one time step instead of the entire time series